+++ src/main/java/org/realityforge/ssf/SimpleSessionInfo.java
@@ -9,15 +9,20 @@ import javax.annotation.Nullable;
 public class SimpleSessionInfo
   implements SessionInfo, Serializable
 {
-  private final String _sessionID;
+  private String _sessionID;
   private long _createdAt;
   private long _lastAccessedAt;
   private Map<String, Serializable> _attributes = new HashMap<>();
 
   public SimpleSessionInfo( @Nonnull final String sessionID )
   {
-    _sessionID = sessionID;
+    setSessionID( sessionID );
     _createdAt = _lastAccessedAt = System.currentTimeMillis();
+  }
+
+  protected void setSessionID( final String sessionID )
+  {
+    _sessionID = sessionID;
   }
 
   @Nonnull