+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 0.2:
+* Rework SimpleSessionInfo to support a mutable sessionID in sub-classes.
 * Rework SessionInfo to support setting of arbitrary attributes. Remove accessor for username and treat
   username as a standard attribute.